check for non-empty shortMsg
The $shortMsg value can be an empty string when the bug number is the last thing in the commit log line: e.g. "[user@php.net] 0df3233344 Fixed bug #1234"

This variable is not currently used anywhere, so we could probably remove it entirely. But for the moment, let's just fix the Notices being printed.